[Examples] Allow overriding of SimpleCppManager identifier
Part of #1202. The upcoming hybrid plugin system will work by locating
multiple plugins that all advertise the same unique identifier and
combining them.

In preparation for e2e testing of this, make the plugin identifier of
SimpleCppManager configurable, so that it can be set to the same
identifier as another plugin on the system.

Signed-off-by: David Feltell <david.feltell@foundry.com>