[Python] Fix terminology `Mapper` initialisation
Discovered whilst working on  #1202. Since the `updateTerminology` API
methods were moved to C++, the input terminology dictionary is no longer
mutated. Instead, the dictionary is copied (by the Python bindings), and
the manager plugin is expected to return a mutated copy.

The `Mapper` class was still assuming the input terminology dict would
be mutated. This was never detected in the unit tests, because we were
erroneously subclassing `Manager` rather than `ManagerInterface`, and so
short-circuiting the Python->C++->Python code path.

So refactor the tests to instead create a mock `ManagerInterface` and
catch this problem. Fix the problem by overwriting the `Mapper`'s
internal terminology dictionary, rather than expecting it to be
mutated.

Signed-off-by: David Feltell <david.feltell@foundry.com>