[Python] Make non-subclassable C++ classes `final`
Discovered whilst working on #1202.

When a Python script defines a class that inherits from a C++ class, the
natural expectation would be that calls to the overridden method in C++
via a C++ base class reference would call out to the Python method
implementation.

However, this is not the case unless we have explicitly enabled it in
the bindings by providing a so-called "trampoline" class.

So to prevent confusion, we add a tag to the Python bindings to declare
the class as "final", so that it is an error at Python import time when
a class inherits from that C++ class.

However, there were several classes that we had never added the
`is_final` tag to.

So add this tag to appropriate Python bindings.

Also add to `CppPluginSystemManagerImplementationFactory`, despite it
defining a trampoline class. This class was never intended to be
derived from by Python classes. Hence remove its unnecessary trampoline
class.

Signed-off-by: David Feltell <david.feltell@foundry.com>