[Core] Add hybrid plugin system
Closes #1202.

OpenAssetIO currently supports two different manager plugin backends in
the core library, Python and C++. We expect more backends to be added in
the future. For example, there has been active investigation into a gRPC
backend.

By "backend" we specifically mean an implementation of the abstract base
class `ManagerImplementationFactoryInterface`. The host application
chooses an implementation, then injects it into a `ManagerFactory`,
which is then responsible for producing `Manager` instances that the
host application can use.

The idea of a hybrid plugin system is to allow multiple backend plugin
providers to be composed, such that API calls can be routed to the most
appropriate plugin. For example, performance-critical API requests could
route to a C++ plugin, whilst less critical requests route to a Python
plugin.

As an added bonus, having a built-in facility to compose multiple
backends reduces boilerplate for hosts that wish to support multiple
plugin system backends.

So add a `HybridPluginSystemManagerImplementationFactory`, which
implements the `ManagerImplementationFactoryInterface` interface, and
takes a list of child `ManagerImplementationFactoryInterface`
implementations during construction. The host application can then
provide an instance of a hybrid factory to their `ManagerFactory`,
taking care of the boilerplate of managing multiple plugin systems.

When a host application attempts to load a plugin with a particular
unique identifier, then all child plugin systems are searched. If a
match is found in a single plugin system, then its manager
implementation is returned directly. If a match is found in multiple
plugin systems, then a proxy manager is constructed that routes calls to
the appropriate manager implementation.

The "appropriate" manager implementation to route to varies. In most
cases, the highest priority implementation that advertises the
capability associated with the API call is chosen (i.e. it is routed
based on the response of the manager(s) to `hasCapability(...)`).

The priority order of implementations corresponds to the order of the
child factories in the list that was provided to the hybrid factory on
construction.

Some API methods don't have an associated capability. For these, routing
is "common sense":

For `initialize`, all implementations are provided with all the
settings. For example, if using a `.toml` OpenAssetIO config file (see
`OPENASSETIO_DEFAULT_CONFIG`), then the settings in that file will be
provided to _all_ the matching manager implementations. This could cause
problems if two plugins use the same settings key but require different
values. Future work may look at ways to disambiguate.

For `identifier` and `displayName` the highest priority manager
implementation is chosen.

For `info` and `settings`, the results of all manager implementations
are merged, with the highest priority manager taking precedence if any
dictionary keys conflict.

For `flushCaches` all manager implementations are flushed.

Signed-off-by: David Feltell <david.feltell@foundry.com>