[Python] Allow API compliance if `initialize`d with unknown settings
Part of #1202. The upcoming hybrid plugin system will allow multiple
manager plugins that advertise the same unique ID to be initialised
together and composed.

If the settings passed to the constituent manager plugins must be
tailored for each one, then significant extra work must be done to
design and implement a way to disambiguate and split up the settings.

A much simpler solution is to allow an arbitrary settings dict to be
passed to a manager, and they can select which settings they are
interested in. This way we can provide a single settings dictionary (as
provided, for example, by the OpenAssetIO `.toml` config file
mechanism), and distribute that single dict to all child plugins of a
hybrid plugin.

The disadvantage is if two plugins require the same key to map to a
different value. Future work can be scheduled to tackle this case, if
it's a problem.

So update the `openassetio.test.manager` API Compliance test suite, used
by manager authors to validate their plugin, to no longer assert that
unrecognized settings trigger an exception. As an aside, the exception
that was expected was a `KeyError`, which already made little sense for
C++.

Signed-off-by: David Feltell <david.feltell@foundry.com>