[Examples] Fix SimpleCppManager override default capabilities
Discovered whilst working on #1202. When capabilities were explicitly
given as a list in the settings, the pre-existing capability set was
not reset, meaning it wasn't possible to "switch off" the default
capabilities.

So `clear()` the capability set when settings are detected that should
override it, before filling up the set again with those settings.

Signed-off-by: David Feltell <david.feltell@foundry.com>