[Build] Remove support for Python 3.7
Part of #1340, but it's been coming for a while.

The trigger for finally doing this is the need for
`os.add_dll_directory` support.

So removed Python 3.7 PyPI distributions, and all mention of it from the
codebase.

The new minimum supported Python version is now 3.9. The older versions
may continue to work for a time, but ths is no longer guaranteed nor
supported.

Signed-off-by: David Feltell <david.feltell@foundry.com>