[Tests] Add missing shared_ptr alias tests
Back in #918 we started ensuring that classes that have reference
semantics (i.e. most objects, because Python) have a set of aliases to
their associated `shared_ptr`s. To enforce this we added tests, but
have not kept the tests up to date as new classes have been added.

So add the missing classes to the test.

Unfortunately there is no easy way to ensure these tests are kept up to
date, other than remembering that they're there (or coming across them
again months later).

Signed-off-by: David Feltell <david.feltell@foundry.com>