[Core] Use `std::set` for trait sets
Closes #1339.

C++ has set operation algorithms. However, these only work for sorted
containers. `std::unordered_set` is not a sorted container, and so
(ironically) cannot use set operations.

Most of our prototyping has been done in Python, where set operations
are a part of the core language, and we have used them extensively to
compare and combine trait sets.

The main benefit of `unordered_set` is the better average-case time
complexity. However, for small sets the difference in performance
between `std::set` and `std::unordered_set` is negligible (and may even
favour `std::set`). In practice, we haven't seen any large 100k+ trait
sets in the wild, which is where we'd expect the performance difference
to show up (see links in GitHub issue for details).

Clearly this is a breaking change. However, we do at least use a
`TraitSet` alias everywhere, so as long as no downstream project is
using `std::unordered_set` directly, or non-generic APIs, then source
compatibility will be maintained.

We use the default pybind11 bindings for trait sets, which means the set
is duplicated as a Python native set - this is the same whether it's
`std::unordered_set` or `std::set`. So the Python core library (and
hence downstream Python projects) would be completely unchanged.

Signed-off-by: David Feltell <david.feltell@foundry.com>