[Python] Remove unused _core package
Closes #1000. The `_core` package contained the old debug logging and
API call auditing implementation. This implementation was based on
Python decorators and thus was Python-only. Work to add auditing
functionality in C++ will begin with #418.

Almost all code has now been migrated to C++ and so the Python auditing
implementation is unused. The exception is the `terminology` helper
module, which was still using the old decorators. However, it seems
rather unhelpful if this piece of minor functionality is the only thing
in OpenAssetIO that is audited. The terminology utilities are also
pending migration to C++ in #999, so there's even less reason to keep a
Python-only implementation around.

Signed-off-by: David Feltell <david.feltell@foundry.com>