[Docs] DR for UI delegation plugin architecture
Part of #1367. In order to determine the necessary changes to the
existing API in order to support UI delegation, we must first decide how
UI delegation implementations should be loaded into an application.

So add a decision record that theorises an implementation based on prior
art in Katana and FnAssetAPI, and contrasts this with the naive
implementation that extends the existing plugin scaffolding.

Signed-off-by: David Feltell <david.feltell@foundry.com>