[Core] Use enum for BatchElementError codes
Closes #1060. In #1056 a new enumeration was added for `Access`
categories, and left as a C-style but C++-namespaced enum.

In #557 we added macros as a way to share enum values between C and C++,
when we could just have easily used enums.

We should be consistent regardless, and probably just use enum for
everything as a cross-language common mechanism.

So add `BatchElementError` error codes as a C-style `enum` to the
`internal::errors` namespace, and use these instead of macros in the
C++ API. Hence remove the macros.

The (woefully incomplete and due for a rethink) C API only used one of
the macro values, and it was trivial, so just (effectively) duplicate
the value.

Signed-off-by: David Feltell <david.feltell@foundry.com>