[CI] Test libstdc++ ABI by Python version
Part of #1352. Add a CI job to ensure that the default libstdc++ ABI
used to build OpenAssetIO varies by Python version.

We test this by building the SimpleCppManager example manager plugin
with an explicit libstdc++ ABI based on the Python version, then attempt
to use it in an e2e test via the simpleResolver command-line tool.

Signed-off-by: David Feltell <david.feltell@foundry.com>