[Build] Default libstdc++ ABI based on Python version
Closes #1388. In the VFX industry, the VFX Reference Platform aims to
keep disparate software by multiple vendors broadly using the same
software stack to aid integration into a pipeline. For Python 3.10+
this implies VFX CY23+. For Linux, CY23+ specifies that software should
be built against the non-deprecated libstdc++ ABI.

So it would be fair for a consumer of OpenAssetIO to expect that the
PyPI wheel for Python 3.10+ would be able to load C++ manager plugins
built against the non-deprecated libstdc++ ABI.

However, we always defaulted to the deprecated ABI, and didn't change
from the default when building Python wheels to upload to PyPI.

So change the initial default value of the CMake option
`OPENASSETIO_GLIBCXX_USE_CXX11_ABI` to be based on the detected Python
version. This will affect all CMake builds, including `setup.py` Python
wheel builds.

A further wrinkle is in the Docker images that the `cibuildwheel` tool
uses to build Linux Python wheels. If we use the manylinux2014 CentOS 7
based image, then `OPENASSETIO_GLIBCXX_USE_CXX11_ABI` has no effect -
the deprecated ABI is hardcoded. The manylinux_2_28 image does not have
this problem, but its glibc version 2.28 goes against the CY22 spec.
Luckily, glibc 2.28 is allowed by the CY23+ (i.e. Python 3.10+) specs.
So conditionally use either manylinux2014 or manylinux_2_28, depending
on the Python version.

This doesn't help if Python support is not enabled. A better solution in
that case will come as part of #1353.

Signed-off-by: David Feltell <david.feltell@foundry.com>