[Build] Remove support for toggling libsdtc++ ABI
Closes #1353. Part of #1389.

When the OpenAssetIO build system and CI was first set up, VFX Reference
Platform CY22 was the latest. It was still CentOS 7 based and specified
that the old libstdc++ ABI should be used. We decided that CY22 would be
our default target.

The closest we could get on GitHub CI was Ubuntu 20.04, which matches
the CY22 GCC version (9), but defaults to the new libstdc++ ABI. And so,
we decided to configure CMake to force the deprecated ABI, unless
explicitly toggled.

However, this is not a valid default for CY23+, which specify that the
new ABI should be used.

The closest we can get to "automatically" determining whether the old
ABI should be used or not, is to use whatever is the system default ABI.
So if someone is building on CentOS 7, it'll be the old ABI, and if
someone is building on Rocky 8, it'll be the new ABI.

Package maintainers can still force the ABI externally using e.g.
`CMAKE_CXX_FLAGS=-D_GLIBCXX_USE_CXX11_ABI=0`. In addition, official CY22
support for OpenAssetIO is due to be removed shortly in #1389. So there
seems little point in supporting an explicit CMake option to choose the
ABI (and having the issue of figuring out a suitable default value).

So remove the `OPENASSETIO_GLIBCXX_USE_CXX11_ABI` CMake option, relying
instead on external configuration (either system default,
`CMAKE_CXX_FLAGS`, or similar).

Don't bother mitigating the fact that we now have wider disparity
between our default CI setup (Ubuntu 20.04 with Python 3.9 and GCC 9)
vs. CY22 (since by default the non-deprecated ABI will now be used). As
mentioned, official support for CY22 is due to be removed, and CI
migrated to more closely match a newer VFX Reference Platform version
(#812).

Hence update the CI bootstrap script for Ubuntu 20.04 to configure the
Conan package manager to use the new ABI, matching the default.

Signed-off-by: David Feltell <david.feltell@foundry.com>