[CI] Disable integration tests on Ubuntu pending downstream updates
Part of #1353. The upcoming removal of the
`OPENASSETIO_GLIBCXX_USE_CXX11_ABI`, with its default value of `OFF`,
means that the new libstdc++ ABI will now be used for OpenAssetIO.
However, downstream projects still expect the old ABI and will fail to
link.

So disable the Ubuntu integration tests pending updates to the
downstream projects.

This commit should be reverted in a subsequent PR.

Signed-off-by: David Feltell <david.feltell@foundry.com>