[C] Fix assignStringView -Wstringop-truncation
Part of #812. Building with GCC 11 triggers the compiler warning
`stringop-truncation` due to our use of `strncpy` truncating the
assumed null terminator from the end of the source `string_view`.

But we deliberately don't assume that strings have a null terminator in
the OpenAssetIO C API (so as not to make assumptions about other
languages that link through C bindings), and instead explicitly track
the length of strings.

So a better option is just to use `memcpy`, which doesn't have the extra
and unnecessary (for us) null-terminator filling behaviour of `strncpy`,
and doesn't trigger the compiler warning.

Signed-off-by: David Feltell <david.feltell@foundry.com>