[Core] Remove library beta version number introspection
Part of #1404. Since OpenAssetIO is moving out of beta into the release
candidate phase, it no longer makes sense to provide introspection
of the "beta" number.

We could reword to "rc" but then rc 1.0 would seem lower than beta 2.0,
unless we expose the "pre-release phase" for introspection too... it all
gets a bit messy and requires proper design.

Instead, take the easy way out and remove introspection of pre-release
version numbers altogether. But keep the pre-release tag in the full
version string, so the information remains available for those that wish
to parse it.

Signed-off-by: David Feltell <david.feltell@foundry.com>