[Build] Downgrade pybind11 version in conanfile.py
Closes #1420. Existing known integrations use pybind11 v2.9. If we test
with a newer version, then we run the risk of using features that are
not available in older versions.

Ideally this wouldn't be a problem, because pybind11 is a private
dependency. However, it becomes a problem when a host application also
uses pybind11, for a couple of reasons.

Firstly, build systems may not like building different components of an
application with different dependency versions (e.g. Conan), even if
those dependencies are private to the component.

Secondly, pybind11 is not a truly private dependency. If a host
application also uses pybind11, then they will interact, especially
through exceptions, which are exported symbols with no ABI namespace to
disambiguate (in current versions of pybind11).

An example of a feature that is not supported in older pybind11 versions
is the native support of `frozenset`, which was used in TraitGen in
OpenAssetIO/OpenAssetIO-TraitGen#55 and triggers an exception during
Python->C++ conversion (see OpenAssetIO/OpenAssetIO-TraitGen#94 to
address this).

Signed-off-by: David Feltell <david.feltell@foundry.com>