[Python] C++->Python exception "raise from" support
Closes #1419. If a C++ exception occurs in a pybind11-bound C++
function, then we attempt to translate the C++ exception to a Python
exception. As part of this we execute Python statements, in-particular
we import `openassetio._openassetio.errors`. Attempting to do this
import whilst the Python error indicator is set causes a fatal error in
Python.

This can happen if e.g. in the call stack of a pybind11-bound function,
some C++ code uses the CPython API, which triggers a Python exception,
but then the C++ code goes on to throw its own exception too.

So handle this case by fetching the pre-existing exception, and setting
the `__cause__` and `__context__` of the newly translated C++->Python
exception.

The code duplicates the logic in `pybind11::raise_from`. We can't use
that function directly, since it only supports simple exceptions that
can be created with `PyErr_SetString` (i.e. doesn't support
`BatchElementError`, whose constructor takes arguments other than a
string `message`).

Move the `import` of `_openassetio.errors` into the `setPyException`
function, ensuring it's called _after_ the `PyErr_Fetch` of the "cause"
exception. This then ensures the error indicator is cleared before
attempting to `import`, which could trigger a fatal CPython error.

Signed-off-by: David Feltell <david.feltell@foundry.com>