[Core] EntityReference associative container support
Closes #573. Entity references used to be simple strings. Now they are a
strong type and lose the conveniences that come with strings.

In particular, they could not be used as keys in associative containers,
in Python nor C++.

So add operators and hashing in both C++ and Python that allow entity
references to be used in containers.

For completeness, add the similar operators that are not technically
needed for container support, but may be useful in other situations. In
particular noting that Python infers partner operators (e.g. `>` from
`<`) whereas C++ does not. So adding explicit support for these to C++
provides parity with Python functionality.

Signed-off-by: David Feltell <david.feltell@foundry.com>