[Core] Support cppcoreguidelines Clang-Tidy checks
In a quest to add static analysis checks for unsafe array access as part
of #1196, the cppcoreguidelines set of Clang-Tidy checks seemed to fit
the bill. However, it turns out that there is only a check for
`std::array` and not for e.g. `std::vector`, which is the main cause for
concern (i.e. callbacks defined in `ManagerConveniences.cpp` that take
a `std::vector` index from an untrusted plugin source).

Despite this, many of the other checks in cppcoreguidelines are
valuable.

So enable the cppcoreguidelines set of checks in the `.clang-tidy`
config, and fix the issues that are flagged.

Whilst doing this, it became apparent that `LoggerInterface` and
`ManagerImplementationFactoryInterface` had pure virtual destructors,
when that is not needed. So `default` them, meaning implementors are not
forced to implement a destructor if they don't need to.

There are some limitations to the cleverness of the cppcoreguidelines
checks that are fixed in later versions, but updating the Clang-Tidy
version used in CI is another issue (#1390).

So suppress cases where the rule of 5 shouldn't apply because the class
is a single defaulted destructor, but the `default` is specified in the
source file, rather than header. This is fixed in later versions.

Also suppress many cases flagged in system headers via use of a macro
from the system header, e.g. the `assert` macro, Catch2 test macros.
This is also fixed in later versions.

Dealing with rule of 5 violations led to some discoveries of classes
with dangerously defined default copy/move operations.

So explicitly delete move and assignment of `TraitsData` instances.
Explicitly delete all copy and move of `FileUrlPathConverter`.
Explicitly delete all copy and move of (private) `Regex` and its
`Regex::Match` helper struct (though allow move of `Match` as a
requirement for returning in a `std::optional`). Adding support for
copy/move to these types may be revisited in the future.

Arbitrarily add some missing `#include`s for headers that are otherwise
brought in via transitive includes, as flagged by latest Clang-Tidy's
`include-cleaner` check (not yet enforced on CI, see note above about
upgrading).

Signed-off-by: David Feltell <david.feltell@foundry.com>