[Docs] Remove Thumbnails Doxygen page
Discovered whilst working on #1091.

The Thumbnails page is very misleading. It details a feature that does
not exist in OpenAssetIO. It was written as part of #86, for an old
version of OpenAssetIO that was to be redesigned. This page received
some updates to sort-of reflect the current design, but taken as a whole
is still misleading.

If such a document should exist, it shouldn't be in the
industry-agnostic OpenAssetIO docs, but instead in the MediaCreation
docs.

There is a similar thumbnail example in `Examples.dox`, but makes it
much clearer that it's speculative. Note that even `Examples.dox` is due
to be removed in favour of links to docs in the MediaCreation repository
(see #846).

Signed-off-by: David Feltell <david.feltell@foundry.com>