[Docs] Remove `@needsref` Doxygen tags
Part of #1091. We use a custom `@needsref` Doxygen tag that evaluates to
nothing, and just serves as a convenient tag to grep. This tag was
placed wherever we wanted to refer to something that is not yet
implemented and/or documented, primarily used during the migration from
Python to C++.

All references are now either available, or will never be available, so
we can remove `@needsref` and potentially replace with proper
references.

The one missing reference was for Trait View classes, so add a short
glossary item explaining what these are, and linking to TraitGen and
MediaCreation for further info.

Signed-off-by: David Feltell <david.feltell@foundry.com>