[Tests] Remove tests asserting Python->C++ migration
Part of #1091. The `MethodIntrospector` class, and its usage in tests,
was added as part of a TDD process of migrating methods from Python to
C++. That is, the tests would assert that a method was no longer defined
in Python.

This work was completed long ago, and all the Python sources were
entirely removed as part of that, so the tests no longer serve a useful
purpose.

Signed-off-by: David Feltell <david.feltell@foundry.com>