[CI] Update macOS runner to macos-13
Closes #1434. GitHub CI has started rejecting macOS 12, i.e.

> The macOS-12 environment is deprecated, consider switching to
> macOS-13, macOS-14 (macos-latest) or macOS-15. For more details, see
> actions/runner-images#10721

VFX Reference Platform CY2024 only specifies a Minimum Deployment Target
of 11.0 (CY2025 specifies 12.0). So we cannot match our GitHub runner to
VFX CY2024. Nearest we can do is target the lowest available macOS
version.

So update to `macos-13` runners.

The (default) available compiler on these is apple-clang 15, which isn't
an available option in Conan 1.59 ("ERROR: Invalid setting '15' is not a
valid 'settings.compiler.version' value"), so bump the version of Conan
to the latest v1 at time of writing, i.e. 1.66.

Signed-off-by: David Feltell <david.feltell@foundry.com>