[CI] Bump actions/upload-artifact to v4
`actions/upload-artifact` v3 is now unavailable on GitHub CI.

We had previously skipped a dependabot update due to NodeJS version
requirement conflict with the old VFX22 Docker image, see #1232. Since
we're now testing on VFX24 this shouldn't be a problem.

We also have a problem due to a change in behaviour in v4 (see
https://github.com/actions/upload-artifact?tab=readme-ov-file#breaking-changes)

> Due to how Artifacts are created in this new version, it is no longer
  possible to upload to the same named Artifact multiple times. You must
  either split the uploads into multiple Artifacts with different names,
  or only upload once. Otherwise you will encounter an error.

We must have a unique name for the Python wheels artifact built for
each platform. They will no longer be combined into a single archive.

The uploaded wheels are occasionally used in downstream projects that
must target `main` (temporarily), rather than getting OpenAssetIO from
PyPI, as well as for the PyPI release process.

This change is unfortunate, since downloading a single archive allowed
us to `pip install` the correct build for a platform quite easily. It
also simplified the release process uploading to PyPI.

However, this looks to be mitigated by the latest version of
dawidd6/action-download-artifact, where we can merge multiple archives
that match a regex such that they're all extracted to a single
directory.

Signed-off-by: David Feltell <david.feltell@foundry.com>