squash: Extract ManagerFactory implementation
Create a new `private` directory where common code can go, and extract
the implementation details of `ManagerFactory` into it, so that the
same functionality can be used in the `UIDelegateFactory` class.

We (unfortunately) have precedent in the codebase for various names for
internal implementation detail namespaces/directories - we use `detail`,
`internal` and `private`. The most similar use-case is in
`openassetio-python` where we have a `private` include directory for use
by multiple modules. So use a similar scheme here. Note that we cannot
use `private` as a namespace, since that is a reserved word in C++.

Signed-off-by: David Feltell <david.feltell@foundry.com>

Signed-off-by: David Feltell <david.feltell@foundry.com>