[Python] Add "isPython" to HostInterface.info()
Part of #1445. OpenAssetIO is a mostly seamless dual-language library,
but for some access patterns it may be necessary for managers to be
aware that the host is written in Python to ensure correct behaviour.

For example, UI delegation may require UI elements to be passed to the
host as Python object wrappers around native UI elements.

Signed-off-by: David Feltell <david.feltell@foundry.com>