[Docs] Add new type contribution checklist
Part of #1445. When adding a new type to the OpenAssetIO API it can be
easy to miss one or more of the many places scattered throughout the
codebase that must be updated to properly support and test the type,
even for devs with in-depth knowledge of the codebase.

So add a brief checklist to CODING_STANDARDS that can act as a memory
aid. Don't go into much detail on what each bullet point means -
explanations are located elsewhere, either within that same document or
in the codebase.

Signed-off-by: David Feltell <david.feltell@foundry.com>