[Docs] Remove docs for UnhandledException
Discovered during #1445. This exception type seems to have slipped the
net. It was designed for #1072, to be used in middleware classes such as
`Manager`, whereby exceptions that are not derived from
`OpenAssetIOException` and are thrown from a plugin's method would be
caught and transformed into an `UnhandledException` and re-thrown. This
was to allow hosts to detect exceptions that are caused by its
OpenAssetIO integration.

However, that mechanism was never added, yet `UnhandledException` was
added, with a docstring making false claims.

Removing the exception type is a breaking change, as is adding the
functionality it was designed for (not to mention it's a lot of work for
a potentially undesirable feature). So for now simply blank the
erroneous docstring.

Signed-off-by: David Feltell <david.feltell@foundry.com>