squash: Add uiPolicy
Add a `uiPolicy` method, mirroring the `managementPolicy` method of
manager plugins. This method can be used by the host to determine
ahead of time whether a UI delegation request via `populateUI` can
possibly succeed.

Note that it is necessary, but not sufficient, for `uiPolicy` to return
positively in order for a `populateUI` call to succeed. The `populateUI`
request may still be rejected, based on a more fine-grained analysis of
the request at the time.

It is by checking for a particular policy trait in the response from
`uiPolicy` (typically a `Managed` trait) that the host can determine if
a particular kind of UI delegation request is supported at all. However,
other policy-specific traits are possible too. For example, a
`DisplayName` trait could be imbued with a property providing the tab or
window title that should be used for containers of such UI elements.

In general, policy traits (and their associated properties) provide the
"global" metadata for a particular kind of request for UI delegation,
i.e. the metadata that is not specific to an individual request.

Signed-off-by: David Feltell <david.feltell@foundry.com>