[CI] Add a suppression for GNU Make web link fail
For a few weeks now the `https://www.gnu.org/software/make` link in
`docs/doxygen/README.md` frequently fails the `markdown-link-check` CI
job with

```
 ERROR: 1 dead links found!
[✖] https://www.gnu.org/software/make/ → Status: 0
```

That link invariably works when tried locally, yet fails most (but not
all) of the time when `markdown-link-check` tries it on a GitHub runner.

So simply add a suppression for `https://www.gnu.org`.

Signed-off-by: David Feltell <david.feltell@foundry.com>