[Core] Warn on discovering duplicate plugins
Part of #1445. As part of refactoring the plugin systems, take the
opportunity to bump the log level that warns of duplicate plugins
being discovered.

If two plugins are discovered at different locations but with the same
unique identifier, then chances are something is awry and the user would
want to know about it. So bump the log level to Warning, so that it's
likely to be printed in a default log configuration.

Signed-off-by: David Feltell <david.feltell@foundry.com>